Ensure fmt.Fscan* methods are buffered
If the fmt.Fscan* family of methods are invoked on unbuffered
input, they will consume an extra byte of data. The tests
didn't catch this because they coincidentally used a buffered
input instead of a network connection. Because network
connections are normally unbuffered, the protocol was unable
to read messages from the network without this change. The
tests should also be modified to use a network connection
in at least one test to ensure that messages can be read
properly.