Incorporate grove as storage backend
This has some inefficiencies right now, especially on operations
like leaves_of that make us iterate all of the available nodes.
However, this can be improved with caching, smarter deserialization,
parallelization, and node partitioning.

It has a long way to go, but I think this version of the relay
supports the needed basic operations to act as a server!