Make relay fetch all peer known content at startup
This is a gross way to warm an empty or semi-empty local cache
by brute-force. We just ask the peer for everything they have
(within some limits). Because the API for sending sprout
messages doesn't currently allow you to block until your
message has a response, we have to do some gross time-based
crap. We'll remove this ASAP.